move DefaultSerializer to a separate file, add tests
