+++ lib/active_model_serializers.rb
@@ -3,6 +3,7 @@ require "active_support/core_ext/string/inflections"
 require "active_support/notifications"
 require "active_model"
 require "active_model/array_serializer"
+require "active_model/default_serializer"
 require "active_model/serializer"
 require "active_model/serializer/associations"
 require "set"
